Break up fetch and store macros
