+++ xmtp_mls/src/lib.rs
@@ -36,55 +36,6 @@ pub trait Delete<Model> {
     fn delete(&mut self, key: Self::Key) -> Result<usize, StorageError>;
 }
 
-#[macro_export]
-macro_rules! impl_fetch_and_store {
-    ($model:ty, $table:ident) => {
-        impl $crate::Store<$crate::storage::encrypted_store::DbConnection> for $model {
-            fn store(
-                &self,
-                into: &mut $crate::storage::encrypted_store::DbConnection,
-            ) -> Result<(), $crate::StorageError> {
-                diesel::insert_into($table::table)
-                    .values(self)
-                    .execute(into)
-                    .map_err(|e| $crate::StorageError::from(e))?;
-                Ok(())
-            }
-        }
-
-        impl $crate::Fetch<$model> for $crate::storage::encrypted_store::DbConnection {
-            type Key = ();
-            fn fetch(&mut self, _key: Self::Key) -> Result<Option<$model>, $crate::StorageError> {
-                use $crate::storage::encrypted_store::schema::$table::dsl::*;
-                Ok($table.first(self).optional()?)
-            }
-        }
-    };
-
-    ($model:ty, $table:ident, $key:ty) => {
-        impl $crate::Store<$crate::storage::encrypted_store::DbConnection> for $model {
-            fn store(
-                &self,
-                into: &mut $crate::storage::encrypted_store::DbConnection,
-            ) -> Result<(), $crate::StorageError> {
-                diesel::insert_into($table::table)
-                    .values(self)
-                    .execute(into)
-                    .map_err(|e| $crate::StorageError::from(e))?;
-                Ok(())
-            }
-        }
-
-        impl $crate::Fetch<$model> for $crate::storage::encrypted_store::DbConnection {
-            type Key = $key;
-            fn fetch(&mut self, key: Self::Key) -> Result<Option<$model>, $crate::StorageError> {
-                use $crate::storage::encrypted_store::schema::$table::dsl::*;
-                Ok($table.find(key).first(self).optional()?)
-            }
-        }
-    };
-}
-
 #[cfg(test)]
 mod tests {
     use std::sync::Once;
