+++ xmtp_mls/src/storage/encrypted_store/identity.rs
@@ -3,7 +3,7 @@ use diesel::prelude::*;
 use super::schema::identity;
 use crate::{
     identity::Identity,
-    impl_fetch_and_store,
+    impl_fetch, impl_store,
     storage::serialization::{db_deserialize, db_serialize},
 };
 
@@ -18,7 +18,8 @@ pub struct StoredIdentity {
     rowid: Option<i32>,
 }
 
-impl_fetch_and_store!(StoredIdentity, identity);
+impl_fetch!(StoredIdentity, identity);
+impl_store!(StoredIdentity, identity);
 
 impl StoredIdentity {
     pub fn new(
